+++ RockLib.Logging/DependencyInjection/LoggerBuilder.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Options;
 using RockLib.Configuration;
 using RockLib.Configuration.ObjectFactory;
 using RockLib.Logging.LogProcessing;
+using RockLib.Logging.LogProviders;
 using System;
 using System.Linq;
 
@@ -109,9 +110,9 @@ namespace RockLib.Logging.DependencyInjection
             var contextProviders = options.ContextProviderRegistrations.Select(createContextProvider => createContextProvider(serviceProvider)).ToArray();
 
             if (optionsMonitor != null && options.ReloadOnChange)
-                return new ReloadingLogger(logProcessor, LoggerName, logProviders, contextProviders, optionsMonitor, options, ConfigureOptions);
+                return new ReloadingLogger(serviceProvider, logProcessor, LoggerName, logProviders, contextProviders, optionsMonitor, options, ConfigureOptions);
 
-            return new Logger(logProcessor, LoggerName, options.Level.GetValueOrDefault(),
+            return new Logger(logProcessor, LoggerName, options.Level.GetValueOrDefault(), serviceProvider.GetService<ILogLevelResolver>(),
                 logProviders, options.IsDisabled.GetValueOrDefault(), contextProviders);
         }